Error dialog: Do not require a window refference
Instead use `gio::Application::get_default`, cast it to a
gtk::Application and call ::get_active_window method.

This removes the need of method to have acess to the App struct
or a window refference. Since gtk defaults to single window apps
and this is a simple app there shouldn't be any problem with this
approach, while it will allow further decoupling of the codebase.

Further docs, https://wiki.gnome.org/HowDoI/GtkApplication